requirements.txt
asgiref==3.7.2
coverage==7.2.7
Django==4.2.4
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
